+++ noodles-sam/src/reader/record.rs
@@ -10,9 +10,9 @@ mod sequence;
 mod template_length;
 
 pub(crate) use self::{
-    cigar::parse_cigar, data::parse_data, flags::parse_flags,
-    mapping_quality::parse_mapping_quality, position::parse_alignment_start,
-    read_name::parse_read_name, sequence::parse_sequence, template_length::parse_template_length,
+    cigar::parse_cigar, flags::parse_flags, mapping_quality::parse_mapping_quality,
+    position::parse_alignment_start, read_name::parse_read_name, sequence::parse_sequence,
+    template_length::parse_template_length,
 };
 
 use std::{
@@ -21,7 +21,8 @@ use std::{
 };
 
 use self::{
-    quality_scores::parse_quality_scores, reference_sequence_id::parse_reference_sequence_id,
+    data::parse_data, quality_scores::parse_quality_scores,
+    reference_sequence_id::parse_reference_sequence_id,
 };
 use super::read_line;
 use crate::{alignment::Record, Header};